pretty good collisions

+++ balls.h
@@ -8,7 +8,7 @@ enum {
 };
 
 typedef struct {
-	int x, y;
+	double x, y;
 } Vec;
 
 typedef struct {
@@ -18,17 +18,17 @@ typedef struct {
 typedef struct {
 	Point p; /* position */
 	Vec v; /* velocity */
-	int m; /* mass */
+	double m; /* mass */
 } Ball;
 
 Vec vsub(Vec v1, Vec v2);
-Vec vmuls(Vec v, int a);
-Vec vdivs(Vec v, int a);
-int vdot(Vec v1, Vec v2);
-int vlen(Vec v);
+Vec vmuls(Vec v, double a);
+Vec vdivs(Vec v, double a);
+double vdot(Vec v1, Vec v2);
+double vlen(Vec v);
 Vec unitnorm(Vec v);
 Point ptaddv(Point p, Vec v);
-Vec V(int x, int y);
+Vec V(double x, double y);
 Vec Vpt(Point p, Point q);
 
 int iscollision(Point p, Point q);
